screenshots_through_tor/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
idna==2.10
image==1.5.32
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
requests-file==1.5.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
tldextract==2.2.3
urllib3==1.25.10
